// src/features/types.ts
export interface sidebarLink {
    id: number
    title: string
    path: string
    allowedRoles: string[]
}